Updated Cmp Config + Typst Setup + NVUI Statusline Modified
Updated cmp config to only confirm on <CR> when selected, made selection
initially nothing (requires <TAB>), and cut out a lot of binds.

Setup Typst files with treesitter, lsp, cmp, and spellcheck. Also
installed a plugin to preview them in real time via web browser and
added a mapping for simply converting to PDF.

Updated NVUI statusline to include a '+' next to the name of the current
file if it has been modified. This was a pain in the ass.

Fixed some deprecation on 'scripts/bclose.lua'.